backend/requirements.txt
fastapi
uvicorn
google-genai
sqlalchemy
asyncpg
pydantic
pydantic-settings
cryptography
spacy
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
httpx